Customization Agent/tenant_builder_app/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
langgraph==0.2.39
python-multipart==0.0.9
Pillow==10.4.0